refactored views.py
